Update patch set 25
Patch Set 25:

> Patch Set 24:
> 
> See https://github.com/tc39/proposal-private-methods/pull/52 for the editorial change to phrase private methods in terms of brand checks rather than nonwritable private fields.

If I'm reading this right, this goes back to Toon's alternate suggestion of putting the methods on the context? 

This PR seems to alleviate the (non) concern that new JSFunctions had to be created for each private method, but that was already clear from the discussion in https://github.com/tc39/proposal-private-methods/issues/31

There's still overhead here to store the brands themselves which was the crux of our discussion. I agree that not having to be concerned about the JSFunction's lifetime is useful, but I'm not sure how this is saving memory

Patch-set: 25
Reviewer: Gerrit User 1158954 <1158954@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>