Update patch set 25
Patch Set 25:

> If I'm reading this right, this goes back to Toon's alternate suggestion of putting the methods on the context? 

Yes.

> There's still overhead here to store the brands themselves which was the crux of our discussion. I agree that not having to be concerned about the JSFunction's lifetime is useful, but I'm not sure how this is saving memory

This CL is currently still relying on the existing optimization to make sure that the pointers to the JSFcuntions are stored directly in the descriptor arrays instead of in the objects as fields. For simple methods, I think a possible optimization could be to skip the private symbols and associate the variable directly with the closures in the context slots, so instead of

LdaImmutableCurrentContextSlot [4]  // Load private symbol
ThrowReferenceErrorIfHole [0]
LdaKeyedProperty <this>, [0]  // Load method
Star r0
CallProperty0 r0, <this>, [2]

It would look like

LdaImmutableCurrentContextSlot [4]  // Load method
ThrowReferenceErrorIfHole [0]
Star r0
CallProperty0 r0, <this>, [2]

though for accessors I am not yet sure if it's possible to emit similar bytecode for load/assignments like this, unless it's possible to store accessor pairs in the context?

Patch-set: 25
Reviewer: Gerrit User 1288325 <1288325@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0