Update patch set 19
Patch Set 19:

(8 comments)

With the doc in hand I better understand this. I think most of my concerns at this point have to do with finding a way to name things...previously, much of the weirdness was embedded in the classifier itself, or in interactions with it. The new approach forces the parser to be explicit in what it's doing; the three types of AssignmentExpression parsing methods is the main place where this rears its head.

Patch-set: 19
Reviewer: Gerrit User 1115851 <1115851@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>