[compiler] Add fast path for simple parallel moves
Most parallel do not actually require the heavy analysis in the
GapResolver since there is no overlap between source and destination
registers. Handle these casts by a fast-path to avoid the quadratic
behaviour in the general case.
The fast path is taken for about 2/3 of the non-empty parallel moves on
wasm workloads.

R=herhut@chromium.org, titzer@chromium.org

Bug: v8:8423
Change-Id: I026256fd6cf268f176fccc97f0f427e07629aeb3