MIPS[64] Fix floating point register allocation
Failure to make snapshot in debug mode on MIPS was caused by bad
floating point register allocation for function parameters. According to
MIPS convention, floating point parameters could not be stored on even
numbered floating point register. Which was the case (f7 was allocated).
This patch ensures that f12 and f14 are used for storing fp parameters.

Change-Id: I7d723e03385f756114df60e33798400e97d99d8f