Update patch set 4
Patch Set 4:

(3 comments)

Nice, thanks!

The commit message is a bit misleading. IIUC this is about parallel moves where the source and destination kinds are not the same. So all are loads from stack or all are stores to the stack. If we have register to register moves, then the fast path will not be taken, right?

I don't think we ever have representation changes in gap moves.

Also, for the 2/3 where this hits, how many are single moves? Just curious.

Patch-set: 4
Reviewer: Gerrit User 1265497 <1265497@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>