[api] add optional parameters to abort on uncaught cb
Add two optional parameters to AbortOnUncaughtExceptionCallback: the
message and exception objects. With this, embedders can use information
from the exception/message to decide if the process should abort, as
well as ouput any information relevant from those objects.

For example, when a process crashes on an uncaught rethrown error, the
stack trace printed by V8 when --abort-on-uncaught-exception is set
deviates from the standard behavior: V8 will print the current stack
instead of the error stack. Embedders (such as Node.js) might want to
print the error stack trace as well, and for that they need the message
and exception objects.

R=yangguo@google.com

Change-Id: I8b78d1f5a99f84c3d673f2d30cd7cdb0f0ae5bef