[cpu-profiler] Use unordered_map instead of custom hashmap in sampler
There is no reason to use the custom map here anymore. This let's us
get rid of the custom hash and a lot of casts.

We can also store the SamplerList by value in the map rather than a
pointer, then we don't have to manage the lifetime explicitly.

Also move the SamplerList typedef inside of SamplerManager because it's
an internal detail. Remove the include for <map> because we aren't using
this anywhere anyway.

Change-Id: I787a1b6c3ffc331ec3f36e66d5e07bd115c4cbb4