Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (1 comment)
> 
> Toon, 
> 
> This is to address regressions in this bug: https://bugs.chromium.org/p/chromium/issues/detail?id=923820
> 
> This is fairly mechanical change that changes
> 1. LangaugeMode, ShouldThrow parameters to Maybe<ShouldThrow>.
> 2. Calls GetShouldThrow only when we are about to throw an exception or when we need it for PropertyCallbackArguments. I just left a question for you in ic.cc about PropertyCallbackArguments.CallAccessorSetter.
> 
> I ran few performance try jobs and also ran runtime_stats benchmark. There are no major regressions apart from one JSTest: Proxies/SetIndexWithTraps. I will take a look at it.

I looked into it, it is because the test doesn't return anything from the set method [1]. If I understand correctly the set method is supposed to return true if it succeeds and false if it fails. In strict mode, failures should throw. Since, the set doesn't return any value, we treat it as falsish value and throw in strict mode [2]. If I change the test to return true, then there is no regression. I assume this is a relatively rare case and we can accept that regresssion? Also may be we should fix the test to return true. Since I don't think the test really wanted to test the failure case.

[1] https://cs.chromium.org/chromium/src/v8/test/js-perf-test/Proxies/proxies.js?type=cs&q=test/js-perf-test/Proxies&g=0&l=415

[2] https://cs.chromium.org/chromium/src/v8/src/objects.cc?type=cs&q=JSProxy::SetProperty&g=0&l=6072

> 
> Here's the results for runtime_stats benchmark: https://pinpoint-dot-chromeperf.appspot.com/job/14c7d911540000
> 
> 
> PTAL.
> 
> Thanks,
> Mythri

Patch-set: 5
Reviewer: Gerrit User 1145189 <1145189@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>