[Heap] Don't try and find SFIs FunctionLiteralId when flushing it's bytecode.
For compiled SFIs, finding the FunctionLiteralId is a slow operation that requires
doing a binary search in the script's SFI table. Given the SFI has been flushed,
it is unlikely we will need the FunctionLiteralId again, so just mark the
UncompiledData has having an invalid FunctionLiteralID such that we only do
the search if it is required from the UncompiledData.

This addresses a significant regression on gc_latency when bytecode flushing was
enabled.

BUG=chromium:927038,v8:v8:8395

Change-Id: I5d577b11ab3b835aa4bdcefe8a365199dc2ff8b0