[wasm] Pass async counters to compilation state
The counters are the last use of the Isolate. Remove it by passing in
the counters in a shared_ptr. This way, we can also refactor the
counters later to be per engine or per process.
In a follow-up CL, we can then remove the Isolate, the foreground task
runner and the cancellable task manager from the compilation state.

R=mstarzinger@chromium.org

Bug: v8:8689, v8:8050
Change-Id: I66b4fab77f770cb8a5463244054b428eef1b7c61