Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (12 comments)
> 
> You can easily test UBSan locally with the following in out/ubsan/args.gn:
> 
> is_component_build = false
> is_debug = false
> is_ubsan = true
> symbol_level = 1
> target_cpu = "x64"
> use_goma = false
> v8_enable_test_features = true
> 
> And then:
> autoninja -C out/ubsan d8 cctest unittests inspector-test
> tools/run-tests.py --arch x64 --mode release --outdir=out/ubsan
> 
> Since it's release mode, running the tests is reasonably fast, especially when you already know which subset is most likely to have issues ;-)

Thanks for the tip, but the version of clang that V8 uses (checked out to third_party) fails when it tries to build with ubsan, or i would already be testing with it.

For casting to unsigned, wouldn't that break the math? the base of the lshift can be a negative number, where fdlibm relies on the known behaviour of gcc (and clang)

Patch-set: 1
Reviewer: Gerrit User 1274103 <1274103@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>