[turbofan] Remove MapsParameterInfo
This was a pair of a set of maps and their common instance type (if
any), but it was only used in a printing function. Removing it means one
less heap access during node creations. Eventually we can use the broker
to look at the instance types if we need to.

Bug: v8:7790
Change-Id: If0cadf9b17e3b9e77cffc4f0b69e2585aff7c85c