Merge remote-tracking branch 'origin/master' into vmo_exc
[Fuchsia] Always call zx_vmo_replace_as_executable() in OS::Allocate()

OS::Allocate() previously was calling zx_vmo_replace_as_executable()
when executable access is requested. This breaks the case when the
caller calls OS::SetPermission() later to mark that memory as executable
later. Updated it to call zx_vmo_replace_as_executable() for all VMOs.

This solution is not ideal, but it's consistent with other platforms,
so it's acceptable short-term.

Bug: 8899, chromium:934582
Change-Id: Ibd6367862d35575772a74d41170a886b514b83e8