[Fuchsia] Always call zx_vmo_replace_as_executable() in OS::Allocate()
OS::Allocate() previously was calling zx_vmo_replace_as_executable()
when executable access is requested. This breaks the case when the
OS::SetPermission() is called to mark that memory as executable later.
Updated it to call zx_vmo_replace_as_executable() for all VMOs.

This solution is not ideal, but it's consistent with other platforms,
so it's acceptable short-term.

Bug: 8899, chromium:934582
Change-Id: Ifeb818c93d8b9c80e73a057f5e6f3ca5a7b1c23f