Update patch set 6
Patch Set 6:

Hello,

Here is a patch that allows us to relax range accesses to the mark bitmap. We found that it reduces the overall pause time by roughly 2% when running both octane and the ARES-6 benchmark on Arm64. We've measured the pause time with --trace-gc-nvp and then individual GC events with --enable-tracing.

More specifically, the reason we see improvements with this patch is because of the `Clear()` method. We've essentially turned:

  base::Atomic32* cell_base = reinterpret_cast<base::Atomic32*>(cells());
  for (int i = 0; i < CellsCount(); i++) {   
    base::Relaxed_Store(cell_base + i, 0);
  }
  // This fence prevents re-ordering of publishing stores with the mark-bit
  // clearing stores.
  base:SeqCst_MemoryFence();

Into:

  for (uint32_t i = 0; i < CellCount(); i++) {
    cells()[i] = 0;
  }

Which, on top of removing the barrier, allows clang to replace `Clear()` with `memset()` for the cases where atomicity isn't required. And the same applies to `ClearRange()`.

We were hoping to see improvements in background sweeping and parallel page evacuation but it wasn't clear there was any impact. However, instead we noticed the time spend in the `MC_EVACUATE_PROLOGUE` tracing event was cut in half because we had optimised clearing liveness in the new space with `ResetLinearAllocationArea()` method.

And since this method is also used when starting a scavenge, we would also see a small reduction in the time spend scavenging which we believe explains the overall GC improvement.

What do you think of this micro-optimisation?

Thanks,
Pierre

Patch-set: 6
Reviewer: Gerrit User 1002405 <1002405@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1131764 <1131764@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1183769 <1183769@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false